Split the Materialize settings model and control widgets into their own file
The example file keeps the scene, geometry derivation, and effect
driving; the settings class, the side panel, and the playback bar move
to materialize_settings.dart as reusable widgets with callbacks.
